Make it more minimal

+++ src/arch-install.sh
@@ -2,8 +2,8 @@
 
 trap 'echo Interrupt signal received; exit' SIGINT
 
-# 0 = /var not separate, else separate /var
-BTRFS_Layout="${BTRFS_Layout:-1}"
+# Use Grammak keymap or not. 1=yes, otherwise no.
+Grammak="${Grammak:-1}"
 
 # Detect CPU
 while read VendorID; do
@@ -34,20 +34,22 @@ read Init _ <<< "$(ls -di /proc/1/root/.)"
 
 if (( Root == Init )); then
 
-	# My keymap link
-	URL=https://raw.githubusercontent.com/ides3rt/grammak/master/src/grammak-iso.map
-	File="${URL##*/}"
+	if [[ $Grammak == 1 ]]; then
+		# My keymap link
+		URL=https://raw.githubusercontent.com/ides3rt/grammak/master/src/grammak-iso.map
+		File="${URL##*/}"
 
-	# Download my keymap
-	curl -O "$URL"
-	gzip "$File"
+		# Download my keymap
+		curl -O "$URL"
+		gzip "$File"
 
-	# Set my keymap
-	loadkeys "$File".gz 2>/dev/null
+		# Set my keymap
+		loadkeys "$File".gz 2>/dev/null
 
-	# Remove that file
-	rm -f "$File".gz
-	unset -v URL File
+		# Remove that file
+		rm -f "$File".gz
+		unset -v URL File
+	fi
 
 	# Partition, format, and mount the drive
 	PS3='Select your disk: '
@@ -81,33 +83,8 @@ if (( Root == Init )); then
 		mkdir /mnt/@/usr
 		btrfs su cr /mnt/@/usr/local
 
-		if [[ $BTRFS_Layout == 0 ]]; then
-
-			mkdir -p /mnt/@/var/lib/libvirt
-
-			btrfs su cr /mnt/@/var/cache
-			chattr +C /mnt/@/var/cache
-
-			btrfs su cr /mnt/@/var/lib/flatpak
-
-			btrfs su cr /mnt/@/var/lib/libvirt/images
-			chattr +C /mnt/@/var/lib/libvirt/images
-
-			btrfs su cr /mnt/@/var/local
-			btrfs su cr /mnt/@/var/log
-			btrfs su cr /mnt/@/var/opt
-			btrfs su cr /mnt/@/var/spool
-
-			btrfs su cr /mnt/@/var/tmp
-			chattr +C /mnt/@/var/tmp
-			chmod 1777 /mnt/@/var/tmp
-
-		else
-
-			btrfs su cr /mnt/@/var
-			chattr +C /mnt/@/var
-
-		fi
+		btrfs su cr /mnt/@/var
+		chattr +C /mnt/@/var
 
 		btrfs su cr /mnt/@/.snapshots
 		mkdir /mnt/@/.snapshots/0
@@ -126,40 +103,15 @@ if (( Root == Init )); then
 		mount -o nodev,noatime,compress-force=zstd:1,space_cache=v2,subvol=@/root "$Mapper" /mnt/root
 		mount -o nodev,noatime,compress-force=zstd:1,space_cache=v2,subvol=@/srv "$Mapper" /mnt/srv
 		mount -o nodev,noatime,compress-force=zstd:1,space_cache=v2,subvol=@/usr/local "$Mapper" /mnt/usr/local
-
-		if [[ $BTRFS_Layout == 0 ]]; then
-
-			mkdir -p /mnt/var/{cache,'local',log,opt,spool,tmp}
-			chattr +C /mnt/var/{cache,tmp}
-			chmod 1777 /mnt/var/tmp
-
-			mkdir -p /mnt/var/lib/{flatpak,libvirt/images}
-			chattr +C /mnt/var/lib/libvirt/images
-
-			mount -o nodev,noatime,compress-force=zstd:1,space_cache=v2,subvol=@/var/cache "$Mapper" /mnt/var/cache
-			mount -o nodev,noatime,compress-force=zstd:1,space_cache=v2,subvol=@/var/lib/flatpak "$Mapper" /mnt/var/lib/flatpak
-			mount -o nodev,noatime,compress-force=zstd:1,space_cache=v2,subvol=@/var/lib/libvirt/images "$Mapper" /mnt/var/lib/libvirt/images
-			mount -o nodev,noatime,compress-force=zstd:1,space_cache=v2,subvol=@/var/local "$Mapper" /mnt/var/local
-			mount -o nodev,noatime,compress-force=zstd:1,space_cache=v2,subvol=@/var/log "$Mapper" /mnt/var/log
-			mount -o nodev,noatime,compress-force=zstd:1,space_cache=v2,subvol=@/var/opt "$Mapper" /mnt/var/opt
-			mount -o nodev,noatime,compress-force=zstd:1,space_cache=v2,subvol=@/var/spool "$Mapper" /mnt/var/spool
-			mount -o nodev,noatime,compress-force=zstd:1,space_cache=v2,subvol=@/var/tmp "$Mapper" /mnt/var/tmp
-
-		else
-
-			mount -o nodev,noatime,compress-force=zstd:1,space_cache=v2,subvol=@/var "$Mapper" /mnt/var
-			chattr +C /mnt/var
-
-			mkdir -p /mnt/state/var
-			chattr +C /mnt/state/var
-			mkdir -p /mnt/state/var/lib/pacman
-
-			mkdir -p /mnt/var/lib/pacman
-			mount --bind /mnt/state/var/lib/pacman /mnt/var/lib/pacman
-
-		fi
-
+		mount -o nodev,noatime,compress-force=zstd:1,space_cache=v2,subvol=@/var "$Mapper" /mnt/var
 		mount -o nodev,noatime,compress-force=zstd:1,space_cache=v2,subvol=@/.snapshots "$Mapper" /mnt/.snapshots
+
+		mkdir -p /mnt/state/var
+		chattr +C /mnt/state/var
+		mkdir -p /mnt/state/var/lib/pacman
+
+		mkdir -p /mnt/var/lib/pacman
+		mount --bind /mnt/state/var/lib/pacman /mnt/var/lib/pacman
 
 		unset -v Disk P Mapper
 		break
@@ -176,13 +128,12 @@ if (( Root == Init )); then
 	Args='/^#/d; s/[[:blank:]]+/ /g; s/rw,//; s/,ssd//; s/,subvolid=[[:digit:]]+//'
 	Args+='; s#/@#@#; s#,subvol=@/\.snapshots/0/snapshot##; /\/boot/s/.$/1/'
 	Args+='; s/,codepage=437,iocharset=ascii,shortname=mixed,utf8,errors=remount-ro//'
-	[[ $BTRFS_Layout == 0 ]] || Args+='; /\/var\/lib\/pacman/d'
+	Args+='; /\/var\/lib\/pacman/d'
 	genfstab -U / | sed -E "$Args" | cat -s # > /mnt/etc/fstab
 	unset -v Args
 
-	if [[ $BTRFS_Layout != 0 ]]; then
-		echo '/state/var/lib/pacman /var/lib/pacman none bind 0 0' >> /mnt/etc/fstab
-	fi
+	# Make FSTAB handle bind mount properly
+	echo '/state/var/lib/pacman /var/lib/pacman none bind 0 0' >> /mnt/etc/fstab
 
 	# Optimize FSTAB
 	while read; do
@@ -673,6 +624,9 @@ else
 	# Disallow null password
 	sed -i 's/ nullok//g' /etc/pam.d/system-auth
 
+	# Make store password more secure
+	sed -i '/^password/s/.$/k rounds=65536/' /etc/pam.d/passwd
+
 	# Disable core dump
 	echo '* hard core 0' >> /etc/security/limits.conf
 	sed -i 's/#Storage=external/Storage=none/' /etc/systemd/coredump.conf
@@ -707,28 +661,31 @@ else
 	while :; do passwd "$Username" && break; done
 	unset -v Groups Username GPU
 
-	# My keymap script
-	URL=https://raw.githubusercontent.com/ides3rt/grammak/master/installer.sh
-	File=/tmp/"${URL##*/}"
+	# Specify vconsole.conf(5)
+	VConsole=/etc/vconsole.conf
 
-	# Download my keymap
-	curl -o "$File" "$URL"
-	bash "$File"
+	if [[ $Grammak == 1 ]]; then
+		# My keymap script
+		URL=https://raw.githubusercontent.com/ides3rt/grammak/master/installer.sh
+		File=/tmp/"${URL##*/}"
 
-	# Remove the temp directory
-	rm -rf /grammak
-	unset -v URL File
+		# Download my keymap
+		curl -o "$File" "$URL"
+		bash "$File"
 
-	# My keymap
-	File=/etc/vconsole.conf
-	echo 'KEYMAP=grammak-iso' > "$File"
+		# Remove the temp directory
+		rm -rf /grammak
+		unset -v URL File
+
+		echo 'KEYMAP=grammak-iso' > "$VConsole"
+	fi
 
 	# If Terminus font is installed, then use it
 	if pacman -Q terminus-font &>dev/null; then
-		echo 'FONT=ter-118b' >> "$File"
+		echo 'FONT=ter-118b' >> "$VConsole"
 	fi
 
-	unset -v File
+	unset -v VConsole
 
 	# Remove sudo(8)
 	pacman -Rns --noconfirm sudo
